- Get Auto-Extending buffers ready to support long vectors. - Even though in theory MAX_BUFLENGTH could now be set to R_XLEN_T_MAX (i.e.   2^52), for now we only increase it from 63*2^25 to 2^32 so we can support   buffers of the length of the human genome.
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/S4Vectors@126080 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ inst/include/S4Vectors_defines.h
@@ -47,14 +47,14 @@ struct htab {
  */
 
 typedef struct int_ae {
-	int _buflength;
-	int _nelt;
+	size_t _buflength;
+	size_t _nelt;
 	int *elts;
 } IntAE;
 
 typedef struct int_aeae {
-	int _buflength;
-	int _nelt;
+	size_t _buflength;
+	size_t _nelt;
 	IntAE **elts;
 } IntAEAE; 
 
@@ -64,26 +64,26 @@ typedef struct intpair_ae {
 } IntPairAE;
 
 typedef struct intpair_aeae {
-	int _buflength;
-	int _nelt;
+	size_t _buflength;
+	size_t _nelt;
 	IntPairAE **elts;
 } IntPairAEAE;
 
 typedef struct longlong_ae {
-	int _buflength;
-	int _nelt;
-	long long *elts;
+	size_t _buflength;
+	size_t _nelt;
+	long long int *elts;
 } LLongAE;
 
 typedef struct char_ae {
-	int _buflength;
-	int _nelt;
+	size_t _buflength;
+	size_t _nelt;
 	char *elts;
 } CharAE; 
 
 typedef struct char_aeae {
-	int _buflength;
-	int _nelt;
+	size_t _buflength;
+	size_t _nelt;
 	CharAE **elts;
 } CharAEAE;
